전체조회, 입력 기능 구현1

+++ src/main/java/boot/team/hr/hyun/repo/EmpRepository.java
@@ -3,5 +3,5 @@ package boot.team.hr.hyun.repo;
 import boot.team.hr.hyun.entity.Emp;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmpRepository extends JpaRepository<Emp,Integer> {
+public interface EmpRepository extends JpaRepository<Emp,Long> {
 }
